Huge refactoring Tests updated on Envire branch

+++ kernel3/kernel3.php
@@ -50,8 +50,10 @@ Error_Reporting(F_DEBUG ? E_ALL : 0);
 if (F_DEBUG) 
     ini_set('display_errors', 'On');
 
-if (get_magic_quotes_runtime())
+if (get_magic_quotes_runtime()) {
     set_magic_quotes_runtime(0);
+}
+
 set_time_limit(30);  // not '0' - once i had my script running for a couple of hours collecting GBytes of errors :)
 // here comes the fatal catcher :P
 register_shutdown_function(create_function('', 'if (($a = error_get_last()) && $a[\'type\'] == E_ERROR)
@@ -89,46 +91,52 @@ $base_modules_files = Array(
     F_KERNEL_DIR.DIRECTORY_SEPARATOR.'k3_dbase.php',         // kernel 3 database interface
 );
 // we'll do some trick with caching base modules in one file
-$base_modules_stats = Array();
-foreach ($base_modules_files as $fname)
-    $base_modules_stats[] = filemtime($fname).'|'.filesize($fname);
-$kernel_codecache_dir = is_writable(F_KERNEL_DIR) ? F_KERNEL_DIR : F_CODECACHE_DIR;
-$base_modules_stats = md5(implode('|', $base_modules_stats));
-$base_modules_file = $kernel_codecache_dir.DIRECTORY_SEPARATOR.'.k3.compiled.'.$base_modules_stats;
-if (false && extension_loaded('bcompiler') && file_exists($base_modules_file.'.bc'))
-    require_once($base_modules_file.'.bc');
-elseif (file_exists($base_modules_file.'.php'))
-    require_once($base_modules_file.'.php');
-else
-{
-    foreach (scandir($kernel_codecache_dir) as $fname)
-        if (preg_match('#^.k3\.compiled\.[0-9a-fA-F]{32}\.(php|bc)?$#', $fname))
-            unlink($kernel_codecache_dir.DIRECTORY_SEPARATOR.$fname);
-    $base_modules_eval = array();
-    foreach ($base_modules_files as $fname)
-        $base_modules_eval[] = preg_replace('#^\s*\<\?php\s*|^\s*\<\?\s*|\?\>\s*$#D', '', php_strip_whitespace($fname));
-    $base_modules_eval = preg_replace('#(?<!^)if\s+\(!defined\(\'F_STARTED\'\)\)\s+die\(\'[^\']+\'\)\;\s*#', '', implode(PHP_EOL, $base_modules_eval));
-
-    if (file_put_contents($base_modules_file.'.php', "<?php\n".$base_modules_eval."\n?>"))
-    {
-        if (function_exists('bcompiler_write_file'))
-        {
-            $fileHandle = fopen($base_modules_file.'.bc', 'w');
-            bcompiler_write_header($fileHandle);
-            bcompiler_write_file($fileHandle, $base_modules_file.'.php'); 
-            bcompiler_write_footer($fileHandle);
-            fclose($fileHandle); 
-            unset($fileHandle);
-        }
-        eval($base_modules_eval);
+if (F_DEBUG) {
+    foreach ($base_modules_files as $fname) {
+        require_once($fname);
     }
+} else {
+    $base_modules_stats = Array();
+    foreach ($base_modules_files as $fname)
+        $base_modules_stats[] = filemtime($fname).'|'.filesize($fname);
+    $kernel_codecache_dir = is_writable(F_KERNEL_DIR) ? F_KERNEL_DIR : F_CODECACHE_DIR;
+    $base_modules_stats = md5(implode('|', $base_modules_stats));
+    $base_modules_file = $kernel_codecache_dir.DIRECTORY_SEPARATOR.'.k3.compiled.'.$base_modules_stats;
+    if (false && extension_loaded('bcompiler') && file_exists($base_modules_file.'.bc'))
+        require_once($base_modules_file.'.bc');
+    elseif (file_exists($base_modules_file.'.php'))
+        require_once($base_modules_file.'.php');
     else
     {
-        trigger_error('Kernel: Error creating base modules cache.', E_USER_WARNING);
+        foreach (scandir($kernel_codecache_dir) as $fname)
+            if (preg_match('#^.k3\.compiled\.[0-9a-fA-F]{32}\.(php|bc)?$#', $fname))
+                unlink($kernel_codecache_dir.DIRECTORY_SEPARATOR.$fname);
+        $base_modules_eval = array();
         foreach ($base_modules_files as $fname)
-            require_once($fname);
+            $base_modules_eval[] = preg_replace('#^\s*\<\?php\s*|^\s*\<\?\s*|\?\>\s*$#D', '', php_strip_whitespace($fname));
+        $base_modules_eval = preg_replace('#(?<!^)if\s+\(!defined\(\'F_STARTED\'\)\)\s+die\(\'[^\']+\'\)\;\s*#', '', implode(PHP_EOL, $base_modules_eval));
+
+        if (file_put_contents($base_modules_file.'.php', "<?php\n".$base_modules_eval."\n?>"))
+        {
+            if (function_exists('bcompiler_write_file'))
+            {
+                $fileHandle = fopen($base_modules_file.'.bc', 'w');
+                bcompiler_write_header($fileHandle);
+                bcompiler_write_file($fileHandle, $base_modules_file.'.php');
+                bcompiler_write_footer($fileHandle);
+                fclose($fileHandle);
+                unset($fileHandle);
+            }
+            eval($base_modules_eval);
+        }
+        else
+        {
+            trigger_error('Kernel: Error creating base modules cache.', E_USER_WARNING);
+            foreach ($base_modules_files as $fname)
+                require_once($fname);
+        }
+        unset($base_modules_eval);
     }
-    unset($base_modules_eval);
 }
 unset($kernel_codecache_dir, $base_modules_files, $base_modules_stats, $base_modules_file);
 /**#@+*/
@@ -136,6 +144,24 @@ unset($kernel_codecache_dir, $base_modules_files, $base_modules_stats, $base_mod
 /**
  * the main kernel class
  * used to control all the modules
+ * @property K3_Autoloader $Autoloader
+ * @property FTimer $Timer
+ * @property FCache $Cache
+ * @property FStr $Str
+ * @property K3_Environment $appEnv
+ * @property K3_Request $Request
+ * @property K3_Response $Response
+ * @property K3_Session $Sess
+ * @property K3_Session $Session
+ * @property FLNGData $LNG
+ * @property FDataBase $DBase
+ * @property K3_Registry $Registry
+ *
+ * @property FVISInterface $VIS
+ * @property FFlexyStoreFactory $FlexyStore
+ * @property FCaptcha $Captcha
+ * @property FParser $Parser
+ * @property FMetaFileFactory $MetaFile
  */
 class F extends FEventDispatcher
 {
@@ -173,12 +199,12 @@ class F extends FEventDispatcher
         $this->pool['Timer']      = new FTimer();
         $this->pool['Cache']      = FCache::getInstance();
         $this->pool['Str']        = FStr::getInstance();
+        $this->pool['LNG']        = FLNGData::getInstance();
         $this->pool['appEnv']     = $e = $this->prepareDefaultEnvironment();
         $this->pool['Request']    = $e->getRequest();
         $this->pool['Response']   = $e->getResponse();
         $this->pool['Sess']       =
         $this->pool['Session']    = $e->getSession();
-        $this->pool['LNG']        = FLNGData::getInstance();
         //$this->pool['DBase'] = new FDataBase();
         $this->classes['DBase']    = 'FDataBase';
         $this->classes['Registry'] = 'K3_Registry';
@@ -281,6 +307,7 @@ class F extends FEventDispatcher
 
     /** Handles exceptions and writes logs
      * @access private
+     * @param Exception $e
      * @ignore
      */
     public function handleException(Exception $e)
@@ -303,11 +330,19 @@ class F extends FEventDispatcher
             header ($_SERVER["SERVER_PROTOCOL"].' 503 Service Unavailable');
             header('Content-Type: text/html; charset='.self::INTERNAL_ENCODING);
         }
-        print '<html><head><title>'.$this->LNG->lang('ERR_CRIT_PAGE', false, true).'</title></head><body><h1>'.$this->LNG->lang('ERR_CRIT_PAGE', false, true).'</h1>'.$this->LNG->lang('ERR_CRIT_MESS', false, true).'</body></html>';
+        if (F_DEBUG) {
+            print $eName.': '.$e->getMessage().'. File: '.$e->getFile().'. Line: '.$e->getLine().'.'.PHP_EOL;
+        } else {
+            print '<html><head><title>'.$this->LNG->lang('ERR_CRIT_PAGE', false, true).'</title></head><body><h1>'.$this->LNG->lang('ERR_CRIT_PAGE', false, true).'</h1>'.$this->LNG->lang('ERR_CRIT_MESS', false, true).'</body></html>';
+        }
     }
 
     /** Handles errors, writes logs and throws exceptions for critical errors
      * @access private
+     * @param int $c
+     * @param string $m
+     * @param string $f
+     * @param int $l
      * @ignore
      */
     public function logError($c, $m, $f = '', $l = 0)
@@ -339,7 +374,7 @@ class F extends FEventDispatcher
         elseif (isset($this->classes[$name]) && $this->runModule($name))
             return $this->pool[$name];
 
-        return new FNullObject('F(\''.$name.'\')');
+        return new FNullObject('F()->'.$name);
     }
 
     /** Handles calling modules like functions
@@ -368,4 +403,3 @@ class F extends FEventDispatcher
 function F($name = null) { return F::kernel($name); }
 $GLOBALS['QF'] = $GLOBALS['F'] = F();
 
-?>
